pyproject.toml
[project]
name = "gitastic"
version = "0.1.0"
description = "Ingest Azure DevOps commits into Elasticsearch datastreams"
requires-python = ">=3.12"
dependencies = [
    "elasticsearch>=8.0,<9.0",
    "pyyaml>=6.0",
    "requests>=2.31",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
